requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio>=3.0.0
fuzzywuzzy[speedup]>=0.18.0
plotly>=5.15.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0 